+++ GCR/__init__.py
@@ -1,0 +1,6 @@
+from .base import *
+#from .composite import *
+from .query import *
+from .utils import *
+from .version import __version__
+__author__ = 'Yao-Yuan Mao'
